Customize the crops/poky dockerhub image for our use
This customizes the crops/poky dockerhub image that I have been using to
build the yocto image with some customizations that are needed for
accessing the commercial Qt repositories, and for building boot2qt.

To access commercial Qt repositories, we need to have an ssh key that is
registered with gerrit. I will assume that the docker host user has
this, and has bind-mounted their ~/.ssh directory to /hostuser.ssh in
the container ( `-v ~/.ssh:/hostuser.ssh` ).